+++ Tests/SentryTests/SentryClientTests.swift
@@ -55,6 +55,10 @@ class SentryClientTest: XCTestCase {
             
             transport = TestTransport()
             transportAdapter = TestTransportAdapter(transport: transport, options: options)
+            
+            crashWrapper.internalFreeMemorySize = 123_456
+            crashWrapper.internalAppMemorySize = 234_567
+            crashWrapper.internalFreeStorageSize = 345_678
         }
 
         func getSut(configureOptions: (Options) -> Void = { _ in }) -> SentryClient {
@@ -574,19 +578,29 @@ class SentryClientTest: XCTestCase {
             XCTAssertNil(actual.debugMeta)
         }
     }
-
-    func testCaptureCrash_CrashWrapper_MemoryAndStorage() {
+    
+    func testCaptureCrash_NoExtraContext() {
         let event = TestData.event
-        event.threads = nil
-        event.debugMeta = nil
 
-        fixture.crashWrapper.internalFreeMemorySize = 123_456
-        fixture.crashWrapper.internalAppMemorySize = 234_567
-        fixture.crashWrapper.internalFreeStorageSize = 345_678
         fixture.getSut().captureCrash(event, with: fixture.scope)
 
         assertLastSentEventWithAttachment { actual in
-            let eventFreeMemory = actual.context?["device"]?["free_memory"] as? Int
+            XCTAssertEqual(1, actual.context?["device"]?.count, "The device context should only contain free_memory")
+            
+            let eventFreeMemory = actual.context?["device"]?[SentryDeviceContextFreeMemoryKey] as? Int
+            XCTAssertEqual(eventFreeMemory, 2_000)
+            
+            XCTAssertNil(actual.context?["app"], "The app context should be nil")
+            XCTAssertNil(actual.context?["culture"], "The culture context should be nil")
+        }
+    }
+
+    func testCaptureEvent_AddCurrentMemoryAndStorage() {
+
+        fixture.getSut().capture(event: TestData.event)
+
+        assertLastSentEvent { actual in
+            let eventFreeMemory = actual.context?["device"]?[SentryDeviceContextFreeMemoryKey] as? Int
             XCTAssertEqual(eventFreeMemory, 123_456)
 
             let eventAppMemory = actual.context?["app"]?["app_memory"] as? Int
@@ -596,16 +610,12 @@ class SentryClientTest: XCTestCase {
             XCTAssertEqual(eventFreeStorage, 345_678)
         }
     }
-
-    func testCaptureCrash_DeviceProperties() {
+    
+    func testCaptureEvent_DeviceProperties() {
 #if os(iOS)
-        let event = TestData.event
-        event.threads = nil
-        event.debugMeta = nil
+        fixture.getSut().capture(event: TestData.event)
 
-        fixture.getSut().captureCrash(event, with: fixture.scope)
-
-        assertLastSentEventWithAttachment { actual in
+        assertLastSentEvent { actual in
             let orientation = actual.context?["device"]?["orientation"] as? String
             XCTAssertEqual(orientation, "portrait")
 
@@ -618,18 +628,14 @@ class SentryClientTest: XCTestCase {
 #endif
     }
 
-    func testCaptureCrash_DeviceProperties_OtherValues() {
+    func testCaptureEvent_DeviceProperties_OtherValues() {
 #if os(iOS)
-        let event = TestData.event
-        event.threads = nil
-        event.debugMeta = nil
-
         fixture.deviceWrapper.internalOrientation = .landscapeLeft
         fixture.deviceWrapper.interalBatteryState = .full
 
-        fixture.getSut().captureCrash(event, with: fixture.scope)
+        fixture.getSut().capture(event: TestData.event)
 
-        assertLastSentEventWithAttachment { actual in
+        assertLastSentEvent { actual in
             let orientation = actual.context?["device"]?["orientation"] as? String
             XCTAssertEqual(orientation, "landscape")
 
@@ -639,14 +645,10 @@ class SentryClientTest: XCTestCase {
 #endif
     }
 
-    func testCaptureCrash_Culture() {
-        let event = TestData.event
-        event.threads = nil
-        event.debugMeta = nil
+    func testCaptureEvent_AddCurrentCulture() {
+        fixture.getSut().capture(event: TestData.event)
 
-        fixture.getSut().captureCrash(event, with: fixture.scope)
-
-        assertLastSentEventWithAttachment { actual in
+        assertLastSentEvent { actual in
             let culture = actual.context?["culture"]
             XCTAssertEqual(culture?["calendar"] as? String, "Gregorian Calendar")
             XCTAssertEqual(culture?["display_name"] as? String, "English (United States)")